requirements.txt
pandas>=1.3.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.21.0          # SPLADE model loading (HuggingFace)
sentence-transformers>=2.2.0  # Stage 2 (NQ-Tables)
google-generativeai>=0.3.0    # Stage 3 (OTT-QA) and preprocessing
openai>=1.0.0                  # Stage 3 (NQ-Tables)
requests>=2.28.0               # JINA embeddings REST API (Stage 2 OTT-QA)
scikit-learn>=1.0.0
tqdm>=4.62.0
datasets>=2.0.0

# Additional dependencies for QA evaluation
tiktoken>=0.5.0        # Token counting for OpenAI models
faiss-cpu>=1.7.0       # Dense vector ANN search (Stage 2 index)

# Jupyter notebook support (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing (optional)
pytest>=7.0.0